refactoring script run_wgPlus.py usgin config.conf file 1/2

+++ run_wgPlus.py
@@ -3,72 +3,75 @@ import glob
 import subprocess
 import datetime
 import platform
+import ConfigParser  # Utilise ConfigParser au lieu de configparser dans Python 3
 
-
-webgrabplus_runner = ''
-webgrabplus_dir = ''
+# directories
+wg_dir          = ''
+wg_run_file     = ''
 xmltv_files_dir = ''
-runner_dir = ''
-os_ = platform.system()
+runner_dir      = ''
+out_dir         = ''
+archives_dir    = ''
+
+# credentials
+wg_username     = ''
+wg_email        = ''
+wg_password     = ''
+grab_result     = 0
+
+print(".............................................................................. step 1: 'config.ini'")
+
+os_ = platform.system().lower()
 print 'Os: "' + os_ + '"'
 
+config_dir = 'config_wgPlus_files'
+config_file = '{}/config.conf'.format(config_dir)
 
-# -- CONFIGURATION ---------------------------------------------------------------------------------------------------------------------
-# uncomment or copy this bloc and edit it (only lines with *) according to your configuration, you can fill in the confs of both systems 
-# (windows/linux) if you want your script to be runnable in both without making any changes each time
-# -- -----------------------------------------------------------------------------------------------------------------------------------
-#if os_ == 'Windows':
-#*	webgrabplus_runner = r'c:/path/to/your/wgPlus/install/directory'
-#*	webgrabplus_dir = r'c:/path/to/your/wgPlus/working/directory'
+# check if 'config.ini' exists
+if not os.path.isfile(config_file):
+    print("Erreur : configuration file  '{}' not found.".format(config_file))
+    exit()
 
-#*	xmltv_files_dir = r'c:/path/to/your/xmltv_files/destination'
-#*	runner_dir = r'c:/path/to/this/grab-runner/script'
+# open 'config.ini' file
+config = ConfigParser.ConfigParser()
+config.read(config_file)
 
-#elif os_ == 'Linux':
-#*	webgrabplus_runner = r'/path/to/your/wgPlus/install/directory'
-#*	webgrabplus_dir = r'/path/to/your/wgPlus/working/directory'
+# check if all sections exist
+if not config.has_section("credentials") or not config.has_section(os_) :
+    print("error: file '{}' content is not valid, some sections are empty.".format(config_file))
+    exit()
 
-#*	xmltv_files_dir = r'/path/to/your/xmltv_files/destination'
-#*	runner_dir = r'/path/to/this/grab-runner/script'
+# read credentials
+wg_username = config.get("credentials", "wg_username")
+wg_email = config.get("credentials", "wg_email")
+wg_password = config.get("credentials", "wg_password")
 
-#else:
-#	print 'unknown Os. "' + os_ + '", program is stopped'
-#	exit()
+# check credentials
+if wg_username == '' or wg_email == '' or wg_password == '':
+    print('error: some config elements are empty')
+    exit()
 
-# --------------------------------------------------------------------------------------------------------------------------------------
+# read os directories config
+wg_dir = config.get(os_, "wg_dir")
+wg_run_file = config.get(os_, "wg_run_file")
+xmltv_files_dir = config.get(os_, "xmltv_files_dir")
+runner_dir = config.get(os_, "runner_dir")
 
-if os_ == 'Windows':
-	webgrabplus_runner = r'C:/Program Files (x86)/WebGrab+Plus/bin/WebGrab+Plus.exe'
-	webgrabplus_dir = r'C:/Users/Mbarki/AppData/Local/WebGrab+Plus'
+# check os directories config
+if wg_dir == '' or wg_run_file == '' or xmltv_files_dir == '' or runner_dir == '':
+    print('error: some config({}) elements are empty'.format(os_))
+    exit()
 
-	xmltv_files_dir = r'F:/Developpement/SOURCES/xmltv/xmltv_files'
-	runner_dir = r'F:/Developpement/SOURCES/xmltv/grab-runners'
+out_dir = '{}/out'.format(wg_dir)
+archives_dir = '{}/archives'.format(out_dir)
 
-elif os_ == 'Linux':
-	webgrabplus_runner = r'/run.net.sh'
-	webgrabplus_dir = r'/home/mohammed/dev/epg/grabbers/wgPlus/.wg++'
-
-	xmltv_files_dir = r'/home/mohammed/dev/epg/xmltv_files'
-	runner_dir = r'/home/mohammed/dev/epg/grab-runners'
-else:
-	print 'unknown Os. "' + os_ + '", program is stopped'
-	exit()
-
-print(".............................................................................. step 0: checking ...")
-
-if webgrabplus_runner == '' or webgrabplus_dir == '' or xmltv_files_dir == '' or runner_dir == '':
-	print 'config not found, please edit this script and update the config block, the program is stopped'
-	exit()
-
-out_dir = webgrabplus_dir + r'/out'
-archives_dir = out_dir + r'/archives'
-
-print('webgrabplus_runner : ' + webgrabplus_runner)
-print('webgrabplus_dir    : ' + webgrabplus_dir)
-print('xmltv_files_dir    : ' + xmltv_files_dir)
-print('runner_dir         : ' + runner_dir)
-print('out_dir            : ' + out_dir)
-print('archives_dir       : ' + archives_dir)
+print('Os                  : {}'.format(os_))
+print('wg_run_file         : {}'.format(wg_run_file))
+print('wg_dir              : {}'.format(wg_dir))
+print('xmltv_files_dir     : {}'.format(xmltv_files_dir))
+print('runner_dir          : {}'.format(runner_dir))
+print('out_dir             : {}'.format(out_dir))
+print('archives_dir        : {}'.format(archives_dir))
 
 #exit()
 
@@ -87,27 +90,28 @@ print('resync with Git repository done')
 print(".........................................................................................step 2: init")
 
 print(r'saving "WebGrab++.config.xml" as "WebGrab++_old.config.xml" on WebGrab++ directory ...')
-os.chdir(webgrabplus_dir)
+os.chdir(wg_dir)
 os.system('mv WebGrab++.config.xml WebGrab++.config_old.xml')
 print('done.')
 print('')
 
 os.chdir(runner_dir)
 
-print(r'(force) update runners and configs from Git repository ...')
-os.system('git checkout master')
-os.system('git fetch')
-os.system('git reset --hard origin/master')
-print('done.')
-print('')
+if 1!=1:
+    print(r'(force) update runners and configs from Git repository ...')
+    os.system('git fetch')
+    os.system('git checkout master')
+    os.system('git reset --hard origin/master')
+    print('done.')
+    print('')
 
 print(r'copying "WebGrab++.config.xml" file to WebGrab++ directory ...')
-os.system(r'cp config_wgPlus_files/WebGrab++.config.xml ' + webgrabplus_dir)
+os.system(r'cp config_wgPlus_files/WebGrab++.config.xml ' + wg_dir)
 print('done.')
 print('')
 
 print(r'copying sites.ini files to WebGrab++ directory ...')
-os.system(r'cp config_wgPlus_files/*.ini ' + webgrabplus_dir)
+os.system(r'cp config_wgPlus_files/*.ini ' + wg_dir)
 print('done.')
 print('')
 
@@ -131,37 +135,41 @@ else:
 # call webgrab++:
 print("........................................................................................ step 3: Grabbing")
 
-os.chdir(webgrabplus_dir)
+os.chdir(wg_dir)
 
-if os_ == 'Windows':
-	sortie = subprocess.call(webgrabplus_runner + ' "' + webgrabplus_dir + '"')
-if os_ == 'Linux':
-	sortie = subprocess.call(['.' + webgrabplus_runner, webgrabplus_dir])
+sortie = -1
+if os_ == 'windows':
+    sortie = subprocess.call(wg_run_file + ' "' + wg_dir + '"')
+elif os_ == 'linux':
+    # sortie = subprocess.call(['.' + wg_run_file, wg_dir])
+    sortie = subprocess.call(['. run.net.sh'])
+else:
+    print("Unknown os '{}'".format(os_))
+    sortie = -1
 
-
-exist = glob.glob(out_dir + r'/guide.xmltv')
-
-#exit()
+exit()
 
 # test grabbing result:
-if sortie == 0 and exist:
-    print("Grabbing done ................................................................................... [OK]")
+if sortie == 0 and os.path.isfile(out_dir + r'/guide.xmltv'):
+    print('')
+    print('')
+    print("Grabbing  ..................................................................................... [OK]")
     print('')
 
-    # copy guide.xmltv file:
-    now = datetime.datetime.now()
-    now_str = now.strftime("%d%m%Y_%H_%M")
-    file_name = 'guide_' + now_str + '.xmltv'
+   # copy guide.xmltv file:
+    now_str = datetime.datetime.now().strftime("%d%m%Y_%H_%M")
+    file_name = 'guide_AR_' + now_str + '.xmltv'
 
+    # copy 'guide.xmltv' into files directory:
     os.chdir(out_dir)
-    os.system('cp guide.xmltv ' + file_name)
-
-    print('guide.xmltv copied to "' + file_name)
+    os.system('cp guide.xmltv {}/files/{}'.format(xmltv_files_dir, file_name))
+    os.system('cp guide.xmltv {}/guide_AR.xmltv'.format(xmltv_files_dir))
+    print('guide.xmltv copied to files directory')
     print('')
 
-    # copy file to xmltv_files
-    os.system('cp ' + file_name + ' ' + xmltv_files_dir + '/files')
-    os.system('cp ' + file_name + ' ' + xmltv_files_dir + '/guide_AR.xmltv')
+    print('')
+    print('step 4: done.')
+    print('')
 
     #exit()
 
@@ -171,16 +179,16 @@ if sortie == 0 and exist:
     os.chdir(xmltv_files_dir)
 
     os.system('git add .')
-    os.system('git commit -m "build file: ' + file_name + '"')
+    os.system('git commit -m "build file: {}"'.format(file_name))
     #os.system('git fetch origin')
     #os.system('git rebase origin/master')
     os.system('git push origin master')
 
-    print('file "' + file_name + '" pushed to Git repository')
+    print('file "{}" pushed to Git repository'.format(file_name))
     print('')
 
 else:
-    print("Grabbing error ....................................... [KO]")
+    print("Grabbing error ............................................................................... [KO]")
     # exit()
 
 #  end
